pyproject.toml
[project]
name = "bbc-news-agent"
version = "0.1.0"
description = "BBC News Article Style Analyzer & Generator — multi-agent system that extracts a style guide from BBC articles and generates new ones."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.1",
    "openai>=1.50.0",
    "anthropic>=0.39.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
